move MotionDetector to a more logical package
